requirements.txt
flask
python-dotenv
sqlalchemy
psycopg2-binary
pgvector
requests
pdfplumber
python-docx
azure-identity
azure-storage-blob
openai
numpy
clamd
